Bug 303245 - LocalFile.fetchInfo() returns IFileInfo with wrong size for files > 4GB on Win32

+++ bundles/org.eclipse.core.filesystem/natives/win32/localfile.c
@@ -1,5 +1,5 @@
 /*******************************************************************************
- *  Copyright (c) 2000, 2007 IBM Corporation and others.
+ *  Copyright (c) 2000, 2010 IBM Corporation and others.
  *  All rights reserved. This program and the accompanying materials
  *  are made available under the terms of the Eclipse Public License v1.0
  *  which accompanies this distribution, and is available at
@@ -147,7 +147,7 @@ jboolean convertFindDataToFileInfo(JNIEnv *env, WIN32_FIND_DATA info, jobject fi
     (*env)->CallVoidMethod(env, fileInfo, mid, fileTimeToMillis(info.ftLastWriteTime));
 
 	// file length
-	fileLength =(info.nFileSizeHigh * (MAXDWORD+1)) + info.nFileSizeLow;
+	fileLength =(info.nFileSizeHigh * (((ULONGLONG)MAXDWORD)+1)) + info.nFileSizeLow;
     mid = (*env)->GetMethodID(env, cls, "setLength", "(J)V");
     if (mid == 0) return JNI_FALSE;
     (*env)->CallVoidMethod(env, fileInfo, mid, fileLength);
@@ -214,7 +214,7 @@ jboolean convertFindDataWToFileInfo(JNIEnv *env, WIN32_FIND_DATAW info, jobject 
     (*env)->CallVoidMethod(env, fileInfo, mid, fileTimeToMillis(info.ftLastWriteTime));
 
 	// file length
-	fileLength =(info.nFileSizeHigh * (MAXDWORD+1)) + info.nFileSizeLow;
+	fileLength =(info.nFileSizeHigh * (((ULONGLONG)MAXDWORD)+1)) + info.nFileSizeLow;
     mid = (*env)->GetMethodID(env, cls, "setLength", "(J)V");
     if (mid == 0) return JNI_FALSE;
     (*env)->CallVoidMethod(env, fileInfo, mid, fileLength);
